+++ src/widgets/settingspages/AboutPage.cpp
@@ -4,6 +4,7 @@
 #include "common/Version.hpp"
 #include "util/LayoutCreator.hpp"
 #include "util/RemoveScrollAreaBackground.hpp"
+#include "widgets/BasePopup.hpp"
 #include "widgets/helper/SignalLabel.hpp"
 
 #include <QFormLayout>
@@ -230,15 +231,25 @@ void AboutPage::addLicense(QFormLayout *form, const QString &name,
     auto *a = new QLabel("<a href=\"" + website + "\">" + name + "</a>");
     a->setOpenExternalLinks(true);
     auto *b = new QLabel("<a href=\"" + licenseLink + "\">show license</a>");
-    QObject::connect(b, &QLabel::linkActivated, [licenseLink] {
-        auto *edit = new QTextEdit;
+    QObject::connect(
+        b, &QLabel::linkActivated, [parent = this, name, licenseLink] {
+            auto window =
+                new BasePopup(BaseWindow::Flags::EnableCustomFrame, parent);
+            window->setWindowTitle("Chatterino - License for " + name);
+            window->setAttribute(Qt::WA_DeleteOnClose);
+            auto layout = new QVBoxLayout();
+            auto *edit = new QTextEdit;
 
-        QFile file(licenseLink);
-        file.open(QIODevice::ReadOnly);
-        edit->setText(file.readAll());
-        edit->setReadOnly(true);
-        edit->show();
-    });
+            QFile file(licenseLink);
+            file.open(QIODevice::ReadOnly);
+            edit->setText(file.readAll());
+            edit->setReadOnly(true);
+
+            layout->addWidget(edit);
+
+            window->getLayoutContainer()->setLayout(layout);
+            window->show();
+        });
 
     form->addRow(a, b);
 }